Trekk ut datostreng-generering for paneler. Legg til visning av ferdige fravær
